requirements.txt
# Core LLM
google-generativeai==0.8.3

# LangChain ecosystem - all pinned together
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.7
langchain-core==0.3.19
langgraph==0.2.45

# Vector Store
chromadb==0.5.5

# Document Processing
pypdf==4.3.1

# PII Scrubbing
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6

# API
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9

# Frontend
streamlit==1.38.0

# ML
transformers==4.44.2
scikit-learn==1.5.2
datasets==2.21.0

# Database
sqlalchemy==2.0.35

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
requests==2.32.3

# Testing
pytest==8.3.3